Add monitoring for object-replicator logs disappearing
As noted in lp#1691570, there is an issue with storage I/O during coro-thread cleanup that affects
the swift-object-replicator's ability to complete replication successfully.  This is most easily
witnessed by the lack of the every 5 minute replicated percentage complete messages that should
come from the swift-object-replicator daemon to syslog.  This patch monitors for and alerts on
the condition of the "replicated" line missing from syslog within the past 15 minutes.

Change-Id: Ieb15da3f3f67fa9bcad03151e36c70faae4c36c9
Closes-Bug: 1691570